requirements.txt
tld
cachetools
pyOpenSSL